+++ app/views/channels/_best_channels.html.erb
@@ -1,0 +1,4 @@
+<li class="best-channels">
+  <%= image_tag channel.thumbnail_url, class: "img-circle", size: '30x30' if channel.thumbnail_url %>
+  <%= channel.name %>
+</li>
