+++ app/views/users/_user_profile.html.erb
@@ -16,7 +16,7 @@
   <div class="col-md-6">
     <h3 class="text-center">好きなチャンネルBest3</h3>
     <ol>
-      <%= render partial: 'videos/best_videos', collection: @best_videos, as: 'video' %>
+      <%= render partial: 'channels/best_channels', collection: @best_channels, as: 'channel' %>
     </ol>
   </div>
   <div class="col-md-6">
